+++ src/fsd_entities/home/api/index.ts
@@ -1,0 +1,1 @@
+export * from './get';
